requirements.txt
torch
pdfminer
markdown
transformers
datasets
peft
matplotlib
pandas
scikit-learn
scipy
numpy
openai
openpyxl
google-genai
semanticscholar
arxiv
pydantic>=2.0.0
pydantic-settings>=2.0.0